+++ lib/constants/strings.dart
@@ -14,7 +14,8 @@ const String shareIncomplete =
     'You can\'t share a furdle that hasn\'t been solved yet!';
 const String copiedToClipBoard = 'Score copied to clipboard';
 const String gameAlreadyPlayed = 'You have already played the game';
-const String furdleCracked = 'You cracked the Furdle of the Day!';
+const String furdleCracked =
+    'You cracked the Furdle of the Day!\n\nNext Furdle in';
 const String failedToCrackFurdle =
     'You couldn\'t crack the Furdle of the Day!\nDon\'t let this happen again.\nBetter luck next time!';
 String alphabets = 'abcdefghijklmnopqrstuvwxyz';